# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EpostIn": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "EpostIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdTokenIn": {
        "properties": {
          "id_token": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Id Token",
            "type": "string"
          }
        },
        "required": [
          "id_token"
        ],
        "title": "IdTokenIn",
        "type": "object"
      },
...
      "KodeIn": {
        "properties": {
          "code": {
            "maxLength": 10,
            "minLength": 4,
            "title": "Code",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email",
          "code"
        ],
        "title": "KodeIn",
        "type": "object"
      },
      "KodeSendt": {
        "description": "Svaret fra /email/start. **Alltid 202**, ogsaa for en ukjent adresse - et\nsvar som skilte kjent fra ukjent ville gjort endepunktet til et\noppslagsverk over hvem som bruker appen.\n\nFristene ligger her nettopp for at klienten ikke skal hardkode dem.",
        "properties": {
          "expires_in_minutes": {
            "description": "Levetid for koden.",
            "title": "Expires In Minutes",
            "type": "integer"
          },
          "resend_after_seconds": {
            "description": "Sperrefrist for «send ny kode». Innen fristen svarer endepunktet 429 med Retry-After.",
            "title": "Resend After Seconds",
            "type": "integer"
          },
          "status": {
            "examples": [
              "sendt"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "resend_after_seconds",
          "expires_in_minutes"
        ],
        "title": "KodeSendt",
        "type": "object"
      },
...
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "type": "object"
      },
...
      "TokenPar": {
        "description": "Svaret fra /refresh. Alle andre endepunkter tar `access_token` som\n`Authorization: Bearer <token>`.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "expires_in": {
            "description": "Sekunder til access-tokenet utloeper.",
            "title": "Expires In",
            "type": "integer"
          },
          "public_id": {
            "description": "Den korte ID-en venner soeker opp, f.eks. BF-7Q4K-9F2M.",
            "title": "Public Id",
            "type": "string"
          },
          "refresh_token": {
            "description": "Ugjennomsiktig. Roteres ved hver bruk; to parallelle fornyelser med samme token tolkes som tyveri og tilbakekaller alle oekter.",
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "examples": [
              "Bearer"
            ],
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in",
          "user_id",
          "public_id",
          "display_name"
        ],
        "title": "TokenPar",
        "type": "object"
      },
      "TokenParNyBruker": {
        "description": "Svaret fra de tre innloggingsveiene: tokenparet pluss om kontoen ble\nopprettet av dette kallet.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "expires_in": {
            "description": "Sekunder til access-tokenet utloeper.",
            "title": "Expires In",
            "type": "integer"
          },
          "is_new": {
            "title": "Is New",
            "type": "boolean"
          },
          "public_id": {
            "description": "Den korte ID-en venner soeker opp, f.eks. BF-7Q4K-9F2M.",
            "title": "Public Id",
            "type": "string"
          },
          "refresh_token": {
            "description": "Ugjennomsiktig. Roteres ved hver bruk; to parallelle fornyelser med samme token tolkes som tyveri og tilbakekaller alle oekter.",
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "examples": [
              "Bearer"
            ],
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in",
          "user_id",
          "public_id",
          "display_name",
          "is_new"
        ],
        "title": "TokenParNyBruker",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Bestefar backend",
    "version": "0.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/apple": {
      "post": {
        "operationId": "logg_inn_apple_v1_auth_apple_post",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": false,
            "schema": {
              "default": "",
              "title": "User-Agent",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdTokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenParNyBruker"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logg Inn Apple",
        "tags": [
          "innlogging"
        ]
      }
    },
    "/v1/auth/email/start": {
      "post": {
        "description": "Svarer ALLTID 202, uansett om adressen finnes fra foer. Et svar som skilte\nmellom «kjent» og «ukjent» ville gjort endepunktet til et oppslagsverk\nover hvem som bruker appen.",
        "operationId": "send_kode_v1_auth_email_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpostIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KodeSendt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Kode",
        "tags": [
          "innlogging"
        ]
      }
    },
    "/v1/auth/email/verify": {
      "post": {
        "operationId": "verifiser_kode_v1_auth_email_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": false,
            "schema": {
              "default": "",
              "title": "User-Agent",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenParNyBruker"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verifiser Kode",
        "tags": [
          "innlogging"
        ]
      }
    },
    "/v1/auth/google": {
      "post": {
        "operationId": "logg_inn_google_v1_auth_google_post",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": false,
            "schema": {
              "default": "",
              "title": "User-Agent",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdTokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenParNyBruker"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logg Inn Google",
        "tags": [
          "innlogging"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Idempotent: et ukjent eller allerede utlogget token gir ogsaa 204.\nUtlogging skal aldri feile - da sitter klienten igjen med et token den\ntror er gyldig.",
        "operationId": "logg_ut_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logg Ut",
        "tags": [
          "innlogging"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "forny_v1_auth_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "user-agent",
            "required": false,
            "schema": {
              "default": "",
              "title": "User-Agent",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPar"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forny",
        "tags": [
          "innlogging"
        ]
      }
    },
...
  }
}